feat(web): add radio button to review page
