+++ appeals/web/src/server/appeals/appeal-details/interested-party-comments/view-and-review/view-and-review.router.js
@@ -18,7 +18,7 @@ router
 		validateAppeal,
 		validateComment,
 		validateReviewComment,
-		asyncHandler(controller.renderPostReviewInterestedPartyComment)
+		asyncHandler(controller.postReviewInterestedPartyComment)
 	);
 
 export default router;
